requirements.txt
fastapi
uvicorn
python-multipart
websockets
openai
gtts
python-dotenv
aiofiles
